update menu rekanan. update mitra index update sales index

+++ database/migrations/2023_07_14_155630_create_transaksi_order_detils_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('transaksi_order_id');
             $table->unsignedBigInteger('produk_id');
             $table->integer('jumlah');
-            $table->float('nominal')->nullable();
+            $table->decimal('nominal', 11, 4)->nullable();
             $table->timestamps();
 
             $table->foreign('transaksi_order_id')->references('id')->on('transaksi_orders')->onUpdate('cascade')->onDelete('cascade');
